+++ clientes-app/src/app/detalleproducto/detalleproducto.component.ts
@@ -1,3 +1,4 @@
+import { PersonaServices } from './../persona/persona.service';
 import { ProductoCaracteristicas } from './../producto/productoCaracteristicas.model';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ProductoService } from './../producto/producto.service';
@@ -16,6 +17,8 @@ export class DetalleproductoComponent implements OnInit {
   constructor(
     private productoService: ProductoService,
     private activatedRoute: ActivatedRoute,
+    private personaService: PersonaServices,
+    private router: Router
   ) {}
 
   ngOnInit(): void {
@@ -39,5 +42,15 @@ export class DetalleproductoComponent implements OnInit {
 
   getProducto(): Producto {
     return this.producto;
+  }
+
+  tienePermisos(role: string): boolean {
+    return this.personaService.hasRole(role);
+  }
+
+  EliminarProducto(){
+    return this.productoService.eliminarProducto(this.producto.id).subscribe(
+      (x) =>{this.router.navigate(['productosgrid'])}
+    );
   }
 }