+++ ectemplate-servicio-productos/src/main/java/net/ostemplate/app/productos/models/dao/TiendaRepository.java
@@ -1,0 +1,9 @@
+package net.ostemplate.app.productos.models.dao;
+
+import org.springframework.data.repository.CrudRepository;
+
+import net.ostemplate.app.productos.models.entity.Tienda;
+
+public interface TiendaRepository extends CrudRepository<Tienda, Long> {
+
+}