+++ ectemplate-servicio-productos/src/main/java/net/ostemplate/app/productos/models/entity/Tienda.java
@@ -21,6 +21,15 @@ public class Tienda implements Serializable{
 	@Column(name="id")
 	private Long id;
 	
+
+	public Long getId() {
+		return id;
+	}
+
+	public void setId(Long id) {
+		this.id = id;
+	}
+
 	public String getNombreTienda() {
 		return nombreTienda;
 	}