+++ ectemplate-servicio-productos/src/main/java/net/ostemplate/app/productos/models/service/ImagenProductoI.java
@@ -1,0 +1,8 @@
+package net.ostemplate.app.productos.models.service;
+
+public interface ImagenProductoI {
+
+
+	void borrarImagen(String imagen);
+
+}