+++ ectemplate-servicio-productos/src/main/java/net/ostemplate/app/productos/models/service/PedidoServiceI.java
@@ -13,6 +13,10 @@ public interface PedidoServiceI {
 	Pedido findById(Long id);
 
 	void borraPedido(Long id);
+
+	List<Pedido> recuperaPedidosPorEstado(String estado);
+
+	List<Pedido> recuperaPedidos();
 	
 
 }